Oxford Covid-19 Latest Data as of 10/28/2020
Oxford Covid-19 Government Tracker Response. The CSV file /data/OxCGRT_latest.csv reports country- and state-level data presented in "country-day" format (or "state-day" as the case may be), with a list of all indicators for each country as a single row each day. This CSV is updated every hour from the main database, and the badge above shows whether this data link is functioning correctly.